feat: canvas props config (#1031)
