+++ packages/client/hmi-client/src/temp/tera-simulation-workflow.vue
@@ -1,5 +1,5 @@
 <template>
-	<infinite-canvas>
+	<infinite-canvas debug-mode :last-transform="lastTransform" @save-transform="saveTransform">
 		<template #foreground></template>
 		<template #data>
 			<div style="font-size: 24px; padding: 10px; background: #9ef">This is a DataLayer DIV</div>
@@ -8,5 +8,15 @@
 </template>
 
 <script setup lang="ts">
-import InfiniteCanvas from '@/temp/tera-infinite-canvas.vue';
+import InfiniteCanvas from '@/components/widgets/tera-infinite-canvas.vue';
+
+const lastTransform = {
+	k: 0.75,
+	x: 0,
+	y: 0
+};
+
+function saveTransform(transform: d3.ZoomTransform) {
+	console.log(transform);
+}
 </script>
